fix

See merge request edusoho/edusoho!6264

+++ src/Biz/OrderFacade/Product/CourseProduct.php
@@ -88,18 +88,16 @@ class CourseProduct extends Product implements OrderStatusCallback
     public function onOrderRefundRefunded($orderRefundItem)
     {
         $orderItem = $orderRefundItem['order_item'];
-        $this->getCourseMemberService()->removeStudent($orderItem['target_id'], $orderItem['user_id']);
+        $member = $this->getCourseMemberService()->getCourseMember($orderItem['target_id'], $orderItem['user_id']);
+        if (!empty($member)) {
+            $this->getCourseMemberService()->removeStudent($orderItem['target_id'], $orderItem['user_id']);
+        }
     }
 
     public function onOrderRefundRefused($orderRefundItem)
     {
         $orderItem = $orderRefundItem['order_item'];
         $this->getCourseMemberService()->unlockStudent($orderItem['target_id'], $orderItem['user_id']);
-    }
-
-    public function getOwner($userId)
-    {
-        return $this->getCourseMemberService()->getCourseMember($this->targetId, $userId);
     }
 
     /**
